Differentiate between error and message
Before this change, we didn't differentiate between the error we're
returning and a user-viewable message associated with the response.

Sometimes, we will have a message to display to the user when an error
happens, while, othertimes we won't. By separating these out, we can
separate "internal" and "external" error messages.