Add cleanup feature and resize teardown tests
As a part of continued work inside Manifold's platform, we are beginning
to add rollback and cleanup phases to the provisioning and resize
operations.

During a provision, we may need to delete a resource *before* we
complete the flow (which includes creating a set of credentials). This
was not explicitly tested as a part of Grafton, so we've added a test.
It's also possible for something to wrong after creation of a credential
set has completed, however, this would resemble a normal deprovision
which is already tested.

In the case of resize, it's possible that something could go wrong on
Manifold's side after the resize was successful. At which point,
Manifold would then resize the instance back to the original state.